+++ src/components/settings/UserManagementSection.tsx
@@ -3,9 +3,16 @@ import React, { useState, useEffect } from 'react';
 import { supabase } from '@/integrations/supabase/client';
 import { useToast } from '@/hooks/use-toast';
 import { useAuth } from '@/contexts/AuthContext';
-import { UserCog } from 'lucide-react';
-import UsersTable from './UsersTable';
-import UserInviteDialog from './UserInviteDialog';
+import { Button } from '@/components/ui/button';
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
+import { Loader2, UserCog, UserPlus } from 'lucide-react';
+import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
+import { Input } from '@/components/ui/input';
+import { z } from 'zod';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
 
 type UserRoleOption = 'global_administrator' | 'owner' | 'administrator';
 
@@ -17,11 +24,32 @@ interface UserProfile {
   created_at: string;
 }
 
+const newUserSchema = z.object({
+  email: z.string().email('Please enter a valid email address'),
+  fullName: z.string().min(2, 'Full name must be at least 2 characters'),
+  password: z.string().min(6, 'Password must be at least 6 characters'),
+  role: z.enum(['administrator', 'owner', 'global_administrator'])
+});
+
+type NewUserFormValues = z.infer<typeof newUserSchema>;
+
 const UserManagementSection = () => {
   const [users, setUsers] = useState<UserProfile[]>([]);
   const [loading, setLoading] = useState(true);
+  const [inviteDialogOpen, setInviteDialogOpen] = useState(false);
+  const [creatingUser, setCreatingUser] = useState(false);
   const { toast } = useToast();
   const { user: currentUser, isGlobalAdmin } = useAuth();
+
+  const form = useForm<NewUserFormValues>({
+    resolver: zodResolver(newUserSchema),
+    defaultValues: {
+      email: '',
+      fullName: '',
+      password: '',
+      role: 'administrator'
+    }
+  });
 
   useEffect(() => {
     fetchUsers();
@@ -49,6 +77,85 @@ const UserManagementSection = () => {
     }
   };
 
+  const updateUserRole = async (userId: string, newRole: UserRoleOption) => {
+    try {
+      const { error } = await supabase
+        .from('profiles')
+        .update({ role: newRole })
+        .eq('id', userId);
+
+      if (error) throw error;
+
+      setUsers(users.map(user => 
+        user.id === userId ? { ...user, role: newRole } : user
+      ));
+
+      toast({
+        title: 'Success',
+        description: 'User role updated successfully'
+      });
+    } catch (error) {
+      console.error('Error updating user role:', error);
+      toast({
+        title: 'Error',
+        description: 'Failed to update user role',
+        variant: 'destructive'
+      });
+    }
+  };
+
+  const createNewUser = async (values: NewUserFormValues) => {
+    setCreatingUser(true);
+    try {
+      const { data: authData, error: authError } = await supabase.auth.admin.createUser({
+        email: values.email,
+        password: values.password,
+        email_confirm: true
+      });
+
+      if (authError) throw authError;
+
+      if (authData.user) {
+        const { error: profileError } = await supabase
+          .from('profiles')
+          .update({
+            full_name: values.fullName,
+            role: values.role as UserRoleOption
+          })
+          .eq('id', authData.user.id);
+
+        if (profileError) throw profileError;
+      }
+
+      toast({
+        title: 'Success',
+        description: `User ${values.email} created successfully`
+      });
+
+      form.reset();
+      setInviteDialogOpen(false);
+      await fetchUsers();
+    } catch (error: any) {
+      console.error('Error creating user:', error);
+      toast({
+        title: 'Error',
+        description: error.message || 'Failed to create user',
+        variant: 'destructive'
+      });
+    } finally {
+      setCreatingUser(false);
+    }
+  };
+
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center py-4">
+        <Loader2 className="h-8 w-8 animate-spin text-blue-500" />
+        <span className="ml-2">Loading user data...</span>
+      </div>
+    );
+  }
+
   return (
     <div>
       <div className="flex items-center justify-between mb-4">
@@ -63,17 +170,151 @@ const UserManagementSection = () => {
         </div>
 
         {isGlobalAdmin && (
-          <UserInviteDialog onUserInvited={fetchUsers} />
+          <Dialog open={inviteDialogOpen} onOpenChange={setInviteDialogOpen}>
+            <DialogTrigger asChild>
+              <Button variant="outline" className="flex items-center text-zinc-950">
+                <UserPlus className="mr-2 h-4 w-4" />
+                Add New User
+              </Button>
+            </DialogTrigger>
+            <DialogContent>
+              <DialogHeader>
+                <DialogTitle>Create New User</DialogTitle>
+                <DialogDescription>
+                  Add a new administrator to the platform.
+                </DialogDescription>
+              </DialogHeader>
+
+              <Form {...form}>
+                <form onSubmit={form.handleSubmit(createNewUser)} className="space-y-4">
+                  <FormField
+                    control={form.control}
+                    name="email"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Email</FormLabel>
+                        <FormControl>
+                          <Input {...field} placeholder="user@example.com" />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="fullName"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Full Name</FormLabel>
+                        <FormControl>
+                          <Input {...field} placeholder="John Doe" />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="password"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Password</FormLabel>
+                        <FormControl>
+                          <Input {...field} type="password" placeholder="••••••••" />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="role"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Role</FormLabel>
+                        <Select onValueChange={field.onChange} defaultValue={field.value}>
+                          <FormControl>
+                            <SelectTrigger>
+                              <SelectValue placeholder="Select a role" />
+                            </SelectTrigger>
+                          </FormControl>
+                          <SelectContent>
+                            <SelectItem value="administrator">Administrator</SelectItem>
+                            <SelectItem value="owner">Owner</SelectItem>
+                            <SelectItem value="global_administrator">Global Administrator</SelectItem>
+                          </SelectContent>
+                        </Select>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  <DialogFooter>
+                    <Button type="submit" disabled={creatingUser}>
+                      {creatingUser ? (
+                        <>
+                          <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                          Creating...
+                        </>
+                      ) : 'Create User'}
+                    </Button>
+                  </DialogFooter>
+                </form>
+              </Form>
+            </DialogContent>
+          </Dialog>
         )}
       </div>
       
-      <UsersTable 
-        users={users} 
-        loading={loading} 
-        currentUserId={currentUser?.id} 
-        isGlobalAdmin={isGlobalAdmin}
-        onRefresh={fetchUsers}
-      />
+      <Table>
+        <TableHeader>
+          <TableRow>
+            <TableHead>Name</TableHead>
+            <TableHead>Email</TableHead>
+            <TableHead>Role</TableHead>
+            <TableHead>Created</TableHead>
+            <TableHead>Actions</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {users.map(user => (
+            <TableRow key={user.id}>
+              <TableCell>{user.full_name || 'N/A'}</TableCell>
+              <TableCell>{user.email}</TableCell>
+              <TableCell>
+                {currentUser?.id !== user.id && isGlobalAdmin ? (
+                  <Select
+                    defaultValue={user.role}
+                    onValueChange={(value: UserRoleOption) => updateUserRole(user.id, value)}
+                  >
+                    <SelectTrigger className="w-40">
+                      <SelectValue placeholder="Select role" />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectItem value="administrator">Administrator</SelectItem>
+                      <SelectItem value="owner">Owner</SelectItem>
+                      <SelectItem value="global_administrator">Global Administrator</SelectItem>
+                    </SelectContent>
+                  </Select>
+                ) : (
+                  <span className="capitalize">{user.role.replace('_', ' ')}</span>
+                )}
+              </TableCell>
+              <TableCell>{new Date(user.created_at).toLocaleDateString()}</TableCell>
+              <TableCell>
+                {currentUser?.id !== user.id && (
+                  <Button variant="outline" size="sm" onClick={() => fetchUsers()}>
+                    Refresh
+                  </Button>
+                )}
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
     </div>
   );
 };
