+++ supabase/config.toml
@@ -1,75 +1,76 @@
 
-# A string used to distinguish different Supabase projects on the same host. Defaults to the working
-# directory name when running `supabase init`.
 project_id = "gbypnkiziennhzqbhqtr"
 
 [api]
-# Port to use for the API URL.
+enabled = true
 port = 54321
-# Schemas to expose in your API. Tables, views and stored procedures in this schema will get API
-# endpoints. public and storage are always included.
 schemas = ["public", "storage", "graphql_public"]
-# Extra schemas to add to the search_path of every request. public is always included.
 extra_search_path = ["public", "extensions"]
-# The maximum number of rows returns from a view, table, or stored procedure. Limits payload size
-# for accidental or malicious requests.
 max_rows = 1000
 
 [db]
-# Port to use for the local database URL.
 port = 54322
-# The database major version to use. This has to be the same as your remote database's. Run `SHOW
-# server_version;` on the remote database to check.
+shadow_port = 54320
 major_version = 15
 
-[studio]
-# Port to use for Supabase Studio.
-port = 54323
-
-# Email testing server. Emails sent with the local dev setup are not actually sent - rather, they
-# are monitored, and you can view the emails that would have been sent from the web interface.
-[inbucket]
-# Port to use for the email testing server web interface.
-port = 54324
-smtp_port = 54325
-pop3_port = 54326
+[realtime]
+enabled = true
 
 [storage]
-# The maximum file size allowed (e.g. "5MB", "500KB").
+enabled = true
 file_size_limit = "50MiB"
 
 [auth]
-# The base URL of your website. Used as an allow-list for redirects and for constructing URLs used
-# in emails.
+enabled = true
 site_url = "http://localhost:3000"
-# A list of *exact* URLs that auth providers are permitted to redirect to post authentication.
 additional_redirect_urls = ["https://localhost:3000"]
-# How long tokens are valid for, in seconds. Defaults to 3600 (1 hour), maximum 604,800 seconds (1
-# week).
 jwt_expiry = 3600
-# Allow/disallow new user signups to your project.
+enable_refresh_token_rotation = true
+refresh_token_reuse_interval = 10
 enable_signup = true
 
 [auth.email]
-# Allow/disallow new user signups via email to your project.
 enable_signup = true
-# If enabled, a user will be required to confirm any email change on both the old, and new email
-# addresses. If disabled, only the new email is required to confirm.
 double_confirm_changes = true
-# If enabled, users need to confirm their email address before signing in.
 enable_confirmations = false
 
-[auth.sms]
-# Allow/disallow new user signups via SMS to your project.
-enable_signup = true
-# If enabled, users need to confirm their phone number before signing in.
-enable_confirmations = false
+[auth.external.github]
+enabled = false
+client_id = ""
+secret = ""
+redirect_uri = ""
+url = ""
 
-# Configure Edge Functions
+[analytics]
+enabled = false
+port = 54327
+vector_port = 54328
+backend = "postgres"
+
 [functions]
-# Port to use for Edge Functions.
-port = 54321
-
-# For functions.invite-user
-[functions.invite-user]
-verify_jwt = true
+[functions.process-scheduled-posts]
+verify_jwt = false
+[functions.generate-blog-post]
+verify_jwt = false
+[functions.fact-check-post]
+verify_jwt = false
+[functions.generate-meta-descriptions]
+verify_jwt = false
+[functions.markdown-to-pdf]
+verify_jwt = false
+[functions.revise-blog-post]
+verify_jwt = false
+[functions.generate-topic]
+verify_jwt = false
+[functions.generate-podcast]
+verify_jwt = false
+[functions.generate-podcast-audio]
+verify_jwt = false
+[functions.process-podcast-audio]
+verify_jwt = false
+[functions.create-checkout]
+verify_jwt = false
+[functions.get-subscription]
+verify_jwt = false
+[functions.stripe-webhook]
+verify_jwt = false
