--- supabase/functions/invite-user/index.ts
@@ -1,65 +1,0 @@
-
-import { serve } from "https://deno.land/std@0.168.0/http/server.ts";
-import { createClient } from "https://esm.sh/@supabase/supabase-js@2.38.4";
-import { corsHeaders } from "../_shared/cors.ts";
-
-// Get environment variables
-const supabaseUrl = Deno.env.get('SUPABASE_URL') ?? '';
-const supabaseServiceKey = Deno.env.get('SUPABASE_SERVICE_ROLE_KEY') ?? '';
-
-// Initialize Supabase client with service role key
-const supabase = createClient(supabaseUrl, supabaseServiceKey);
-
-serve(async (req) => {
-  // Handle CORS preflight requests
-  if (req.method === 'OPTIONS') {
-    return new Response(null, { headers: corsHeaders });
-  }
-
-  // Add CORS headers to all responses
-  const headers = { ...corsHeaders, 'Content-Type': 'application/json' };
-
-  try {
-    // Parse request body
-    const { email, fullName, role } = await req.json();
-
-    console.log(`Attempting to invite user: ${email} with role: ${role}`);
-
-    // Validate inputs
-    if (!email || !role) {
-      return new Response(
-        JSON.stringify({ error: 'Email and role are required' }),
-        { status: 400, headers }
-      );
-    }
-
-    // Invite user via email using admin API with service role key
-    const { data, error } = await supabase.auth.admin.inviteUserByEmail(email, {
-      data: {
-        full_name: fullName || '',
-        role: role
-      }
-    });
-
-    if (error) {
-      console.error('Error inviting user:', error);
-      return new Response(
-        JSON.stringify({ error: error.message }),
-        { status: 400, headers }
-      );
-    }
-
-    console.log('User invitation sent successfully');
-    
-    return new Response(
-      JSON.stringify({ success: true, data }),
-      { status: 200, headers }
-    );
-  } catch (error) {
-    console.error('Error in invite-user function:', error);
-    return new Response(
-      JSON.stringify({ error: error.message }),
-      { status: 500, headers }
-    );
-  }
-});
